backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
greenlet==2.0.2
asyncpg==0.29.0
alembic==1.12.1

# Cache & Queue
redis==5.0.1
celery==5.3.4

# AI/ML
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
diffusers==0.24.0
accelerate==0.24.1
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4

# File Storage
minio==7.2.0

# Utilities
pydantic==2.5.0
email-validator==2.3.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
loguru==0.7.2
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
httpx==0.25.2

# WebSocket
websockets==12.0

# Image processing
scikit-image==0.22.0

# Additional utilities
aiofiles==23.2.1